front: Move VirtualRow to the shared components.
Discover's card list is about to be windowed the way the map-mode
list is (#573), and the row that places itself where the virtualiser
says — and registers with measureElement through a stable ref, the
lesson its header comment carries — is the piece both lists need
verbatim. Move it to components/shared rather than importing across
feature directories or forking it.

[Issue: #573]
